The return type of HugeIntWord::get_value() was changed from unsigned long to uint32_t. The types of variables that use the return value were also changed.

+++ HugeIntegers/HugeIntWord.cpp
@@ -15,7 +15,7 @@ HugeIntWord::HugeIntWord(unsigned long value, HugeIntWord* lowerSignificantWord)
     lowerSignificantWord->set_more_significant_word(this);
 }
 
-unsigned long HugeIntWord::get_value() const {
+uint32_t HugeIntWord::get_value() const {
     return this->value;
 }
